refactor(rendering): move shadows code to a single folder
