+++ engine/CMakeLists.txt
@@ -177,17 +177,18 @@ set(CUBOS_ENGINE_SOURCE
 	"src/render/split_screen/split_screen.cpp"
 	"src/render/bloom/plugin.cpp"
 	"src/render/bloom/bloom.cpp"
-	"src/render/shadows/plugin.cpp"
-	"src/render/shadows/caster.cpp"
-	"src/render/shadows/spot_caster.cpp"
-	"src/render/shadows/directional_caster.cpp"
-	"src/render/shadows/point_caster.cpp"
-	"src/render/shadow_atlas/plugin.cpp"
-	"src/render/shadow_atlas/shadow_atlas.cpp"
-	"src/render/shadow_atlas_rasterizer/plugin.cpp"
-	"src/render/shadow_atlas_rasterizer/shadow_atlas_rasterizer.cpp"
-	"src/render/cascaded_shadow_maps/plugin.cpp"
-	"src/render/cascaded_shadow_maps_rasterizer/plugin.cpp"
+	"src/render/shadows/casters/plugin.cpp"
+	"src/render/shadows/casters/caster.cpp"
+	"src/render/shadows/casters/spot_caster.cpp"
+	"src/render/shadows/casters/directional_caster.cpp"
+	"src/render/shadows/casters/point_caster.cpp"
+	"src/render/shadows/atlas/plugin.cpp"
+	"src/render/shadows/atlas/spot_atlas.cpp"
+	"src/render/shadows/atlas/point_atlas.cpp"
+	"src/render/shadows/atlas_rasterizer/plugin.cpp"
+	"src/render/shadows/atlas_rasterizer/atlas_rasterizer.cpp"
+	"src/render/shadows/cascaded/plugin.cpp"
+	"src/render/shadows/cascaded_rasterizer/plugin.cpp"
 
 	"src/tools/settings_inspector/plugin.cpp"
 	"src/tools/selection/plugin.cpp"
