--- engine/src/render/shadows/spot_caster.cpp
@@ -1,12 +1,0 @@
-#include <cubos/core/ecs/reflection.hpp>
-#include <cubos/core/reflection/external/glm.hpp>
-#include <cubos/core/reflection/external/primitives.hpp>
-
-#include <cubos/engine/render/shadows/spot_caster.hpp>
-
-CUBOS_REFLECT_IMPL(cubos::engine::SpotShadowCaster)
-{
-    return core::ecs::TypeBuilder<SpotShadowCaster>("cubos::engine::SpotShadowCaster")
-        .withField("baseSettings", &SpotShadowCaster::baseSettings)
-        .build();
-}
